Add ExchangeAccess module for fetching currency conversion rates
The ExchangeAccess is made up of two components: a cache and a client for querying external apis for rates. The cache is run inside a supervisor and is the first to be queried. If the information is not available there, then the external api is queried and the cache is updated.