Add TTL for keys in cache
The TTL is based on the time until the next refresh of rates for the API, which is once per day. For example, if the api will refresh all of it's rates at midnight UTC and now it is 22:00 UTC, the recently fetched rates will have a TTL of 2 hours until becoming stale and being removed from the cache. Requests after midnight will have to query the external api again.